+++ lib/components/tonal_button.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class TonalButton extends StatelessWidget {
+  final String text;
+  final Color backgroundColor;
+  final VoidCallback onPressed;
+
+   TonalButton({
+    Key? key,
+    required this.text,
+    required this.backgroundColor,
+    required this.onPressed,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(4.0),
+      child: ConstrainedBox(
+        constraints: const BoxConstraints(
+          minWidth: 200
+        ),
+        child: ElevatedButton(
+          onPressed: onPressed,
+          style: ButtonStyle(
+            backgroundColor: MaterialStateProperty.all<Color>(backgroundColor),
+          ),
+          child: 
+            Text(
+              text,
+              style: TextStyle(color: Colors.black.withOpacity(1.0)),
+            ),
+        )
+      )
+    );
+  }
+}
